break generator config into own file
